Extract Random Effects Samples from BCF Model (legacy alias)

Legacy alias for extractRandomEffectSamples(); delegates to it.

Usage

# S3 method for class 'bcfmodel'
getRandomEffectSamples(object, ...)

Arguments

object

Object of type bcfmodel containing draws of a BCF model and associated sampling outputs.

...

Other parameters to be used in random effects extraction

Value

List of random effect samples (see extractRandomEffectSamples()).

Examples

n <- 100
p <- 5
X <- matrix(runif(n*p), ncol = p)
pi_x <- 0.25 + 0.5*X[,1]
Z <- rbinom(n, 1, pi_x)
rfx_group_ids <- sample(1:2, size = n, replace = TRUE)
rfx_basis <- rep(1.0, n)
mu_x <- X[,1]*2
tau_x <- X[,2]*(-1)
y <- mu_x + tau_x*Z + (-2*(rfx_group_ids==1)+2*(rfx_group_ids==2)) + rnorm(n)
bcf_model <- bcf(X_train=X, Z_train=Z, y_train=y, propensity_train=pi_x,
                 rfx_group_ids_train=rfx_group_ids, rfx_basis_train=rfx_basis,
                 num_gfr=0, num_mcmc=10)
rfx_samples <- getRandomEffectSamples(bcf_model)