Sample one iteration of the leaf parameter variance model (only for univariate basis and constant leaf!)

Sample one iteration of the leaf parameter variance model (only for univariate basis and constant leaf!)

Usage

sample_tau_one_iteration(forest_samples, rng, a, b, sample_num)

Arguments

forest_samples

Container of forest samples

rng

C++ random number generator

a

Leaf variance shape parameter

b

Leaf variance scale parameter

sample_num

Sample index