Object used to get / set global parameters and other global model configuration options in the "low-level" stochtree interface

The "low-level" stochtree interface enables a high degreee of sampler customization, in which users employ R wrappers around C++ objects like ForestDataset, Outcome, CppRng, and ForestModel to run the Gibbs sampler of a BART model with custom modifications. GlobalModelConfig allows users to specify / query the global parameters of a model they wish to run.

Value

Global error variance parameter

Public fields

global_error_variance

Global error variance parameter Create a new GlobalModelConfig object.

Methods

Public methods

GlobalModelConfig$new()

GlobalModelConfig$update_global_error_variance()

GlobalModelConfig$get_global_error_variance()

Method new()

Usage

GlobalModelConfig$new(global_error_variance = 1)

Arguments

global_error_variance

Global error variance parameter (default: 1.0)

Returns

A new GlobalModelConfig object.

Method update_global_error_variance()

Update global error variance parameter

Usage

GlobalModelConfig$update_global_error_variance(global_error_variance)

Arguments

global_error_variance

Global error variance parameter

Method get_global_error_variance()

Query global error variance parameter for this GlobalModelConfig object

Usage

GlobalModelConfig$get_global_error_variance()