Random Effects Dataset C++ Wrapper

Dataset used to sample a random effects model. A random effects dataset consists of three matrices / vectors: group labels, bases, and variance weights. Variance weights are optional.

This class is intended for advanced use cases in which users require detailed control of sampling algorithms and data structures. Minimal input validation and error checks are performed – users are responsible for providing the correct inputs. For tutorials on the "proper" usage of the stochtree's advanced workflow, we provide several vignettes at https://stochtree.ai/

Public fields

data_ptr

External pointer to a C++ RandomEffectsDataset class

Methods

Public methods

RandomEffectsDataset$new()

RandomEffectsDataset$update_basis()

RandomEffectsDataset$update_variance_weights()

RandomEffectsDataset$num_observations()

RandomEffectsDataset$num_basis()

RandomEffectsDataset$get_group_labels()

RandomEffectsDataset$get_basis()

RandomEffectsDataset$get_variance_weights()

RandomEffectsDataset$has_group_labels()

RandomEffectsDataset$has_basis()

RandomEffectsDataset$has_variance_weights()

Method new()

Create a new RandomEffectsDataset object.

Usage

RandomEffectsDataset$new(group_labels, basis, variance_weights = NULL)

Arguments

group_labels

Vector of group labels

basis

Matrix of bases used to define the random effects regression (for an intercept-only model, pass an array of ones)

variance_weights

(Optional) Vector of observation-specific variance weights

Returns

A new RandomEffectsDataset object.

Method update_basis()

Update basis matrix in a dataset

Usage

RandomEffectsDataset$update_basis(basis)

Arguments

basis

Updated matrix of bases used to define random slopes / intercepts

Method update_variance_weights()

Update variance_weights in a dataset

Usage

RandomEffectsDataset$update_variance_weights(
  variance_weights,
  exponentiate = F
)

Arguments

variance_weights

Updated vector of variance weights used to define individual variance / case weights

exponentiate

Whether or not input vector should be exponentiated before being written to the RandomEffectsDataset's variance weights. Default: F.

Method num_observations()

Return number of observations in a RandomEffectsDataset object

Usage

RandomEffectsDataset$num_observations()

Returns

Observation count

Method num_basis()

Return dimension of the basis matrix in a RandomEffectsDataset object

Usage

RandomEffectsDataset$num_basis()

Returns

Basis vector count

Method get_group_labels()

Return group labels as an R vector

Usage

RandomEffectsDataset$get_group_labels()

Returns

Group label data

Method get_basis()

Return bases as an R matrix

Usage

RandomEffectsDataset$get_basis()

Returns

Basis data

Method get_variance_weights()

Return variance weights as an R vector

Usage

RandomEffectsDataset$get_variance_weights()

Returns

Variance weight data

Method has_group_labels()

Whether or not a dataset has group label indices

Usage

RandomEffectsDataset$has_group_labels()

Returns

True if group label vector is loaded, false otherwise

Method has_basis()

Whether or not a dataset has a basis matrix

Usage

RandomEffectsDataset$has_basis()

Returns

True if basis matrix is loaded, false otherwise

Method has_variance_weights()

Whether or not a dataset has variance weights

Usage

RandomEffectsDataset$has_variance_weights()

Returns

True if variance weights are loaded, false otherwise