Class that defines a "tracker" for random effects models, most notably storing the data indices available in each group for quicker posterior computation and sampling of random effects terms.

Stores a mapping from every observation to its group index, a mapping from group indices to the training sample observations available in that group, and predictions for each observation.

Public fields

rfx_tracker_ptr

External pointer to a C++ StochTree::RandomEffectsTracker class

Methods

Public methods

RandomEffectsTracker$new()

Method new()

Create a new RandomEffectsTracker object.

Usage

RandomEffectsTracker$new(rfx_group_indices)

Arguments

rfx_group_indices

Integer indices indicating groups used to define random effects

Returns

A new RandomEffectsTracker object.